questionクラスの定義と各質問クラスがQuestionクラスを継承するよう定義。judgeモジュールは不要
