+++ question/physical.rb
@@ -1,12 +1,7 @@
 require "pry"
-require "./module/judge_module"
-require "./module/question_module"
+require_relative "question"
 
-class Physical
-  #モジュールをinclude
-  include Judge
-  include Question
-  attr_reader :title, :questions, :text
+class Physical < Question
 
   PHYSICAL_QUESTIONS = [["フィジカルは強いほうだ", [2,4]],
                         ["細かな動きが得意だ", [3]],
